feat(vhvu-041): move 13 visuals + 7 utils + viz types into @spaarke/visuals
Extract the cleanly-presentational visuals out of the VisualHost PCF into the
new @spaarke/visuals sibling so they become reusable.

MOVED (git rename, byte-identical unless noted):
- 13 components: MetricCard, MetricCardMatrix, BarChart, LineChart, DonutChart,
  GaugeVisual, HorizontalStackedBar, StatusDistributionBar, TrendCard,
  GradeMetricCard, MiniTable, EventDueDateCard, ErrorBoundary.
- 7 utils: chartColors, valueFormatters, gradeUtils, tokenSetColors,
  trendAnalysis, cardConfigResolver, logger.
- All viz types (control/types/index.ts content) -> src/types.

DEPENDENCY INVERSION: TrendDirection now lives in the package types barrel;
TrendCard imports + re-exports it and trendAnalysis imports it from ../types
(util no longer imports a component).

DIRECTIONAL DEVIATION (step-mode directional, CLAUDE.md 8.5): ChartRenderer
and the 3 self-fetch visuals (CalendarVisual, DueDateCard, DueDateCardList)
STAY in the PCF for now — ChartRenderer is a host dispatcher that consumes
webAPI + the self-fetch trio; moving it would drag webAPI into the package
(NFR-05 breach). It moves in VHVU-050/060 after the self-fetch visuals are
inverted to props-in.

VisualHost repointed to the package via relative src paths (mirrors the
AiSummaryPopover pattern; safe because the package is presentational-only with
zero internal deps). Three pervasive utils (logger, cardConfigResolver,
trendAnalysis) kept as thin re-export shims at their old PCF paths; VHVU-060
finishes the full repoint. @spaarke/visuals declared as a file: dep.

VERIFIED: @spaarke/visuals `tsc --noEmit` green + zero host-coupling
(Xrm/webAPI/ComponentFramework only in the doc comment). VisualHost build:prod
green, bundle 761 KiB, still leak-free (PublicClientApplication/SdapClient=0),
cleanGuid intact, footer v1.4.36. ZERO new skew-casts on the 13 moved
components (the @types/react@18 pin from the prior commit did its job).

FOLLOW-UP (VHVU-090 / test-diet): the 6 moved-component test files were
repointed to the package path (keeps 041 from breaking their module
resolution) but have a PRE-EXISTING @testing-library/react TS2305 infra issue
(no screen/fireEvent export) affecting moved + stayed tests alike — to be
reconciled at test-diet, ideally by re-homing them into @spaarke/visuals with
its own jest harness.
