+++ src/client/pcf/VisualHost/control/components/ChartRenderer.tsx
@@ -8,18 +8,18 @@ import * as React from 'react';
 import { makeStyles, tokens, Text } from '@fluentui/react-components';
 import type { IChartDefinition, IAggregatedDataPoint, DrillInteraction, VisualType, IChartData } from '../types';
 import { VisualType as VT } from '../types';
-import { MetricCard } from './MetricCard';
-import { MetricCardMatrix, type MatrixJustification } from './MetricCardMatrix';
-import { BarChart } from './BarChart';
-import { LineChart } from './LineChart';
-import { DonutChart } from './DonutChart';
-import { StatusDistributionBar } from './StatusDistributionBar';
+import { MetricCard } from '../../../../shared/Spaarke.Visuals/src/components/MetricCard';
+import { MetricCardMatrix, type MatrixJustification } from '../../../../shared/Spaarke.Visuals/src/components/MetricCardMatrix';
+import { BarChart } from '../../../../shared/Spaarke.Visuals/src/components/BarChart';
+import { LineChart } from '../../../../shared/Spaarke.Visuals/src/components/LineChart';
+import { DonutChart } from '../../../../shared/Spaarke.Visuals/src/components/DonutChart';
+import { StatusDistributionBar } from '../../../../shared/Spaarke.Visuals/src/components/StatusDistributionBar';
 import { CalendarVisual, type ICalendarEvent } from './CalendarVisual';
-import { MiniTable, type IMiniTableItem, type IMiniTableColumn } from './MiniTable';
+import { MiniTable, type IMiniTableItem, type IMiniTableColumn } from '../../../../shared/Spaarke.Visuals/src/components/MiniTable';
 import { DueDateCardVisual } from './DueDateCard';
 import { DueDateCardListVisual } from './DueDateCardList';
-import { GaugeVisual } from './GaugeVisual';
-import { HorizontalStackedBar } from './HorizontalStackedBar';
+import { GaugeVisual } from '../../../../shared/Spaarke.Visuals/src/components/GaugeVisual';
+import { HorizontalStackedBar } from '../../../../shared/Spaarke.Visuals/src/components/HorizontalStackedBar';
 import { resolveCardConfig } from '../utils/cardConfigResolver';
 import type { IConfigWebApi } from '../services/ConfigurationLoader';
