+++ src/client/pcf/VisualHost/control/components/__tests__/MiniTable.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
 import { FluentProvider, webLightTheme, webDarkTheme } from '@fluentui/react-components';
-import { MiniTable, IMiniTableColumn, IMiniTableItem } from '../MiniTable';
+import { MiniTable, IMiniTableColumn, IMiniTableItem } from '../../../../../shared/Spaarke.Visuals/src/components/MiniTable';
 
 const renderWithTheme = (component: React.ReactElement, theme = webLightTheme) => {
   return render(<FluentProvider theme={theme}>{component}</FluentProvider>);
