+++ src/client/pcf/VisualHost/control/configurations/matterReportCardTrends.ts
@@ -13,7 +13,7 @@
  * with the correct data bindings per performance area.
  */
 
-import type { TrendDirection } from '../components/TrendCard';
+import type { TrendDirection } from '../../../../shared/Spaarke.Visuals/src/components/TrendCard';
 export type { TrendDirection };
 
 export { calculateSlope, getTrendDirection } from '../utils/trendAnalysis';
