+++ src/client/pcf/VisualHost/package.json
@@ -23,7 +23,8 @@
     "build-storybook": "storybook build"
   },
   "dependencies": {
-    "@spaarke/ui-components": "file:../../shared/Spaarke.UI.Components"
+    "@spaarke/ui-components": "file:../../shared/Spaarke.UI.Components",
+    "@spaarke/visuals": "file:../../shared/Spaarke.Visuals"
   },
   "devDependencies": {
     "@azure/msal-browser": "^3.30.0",
