+++ src/client/shared/Spaarke.Visuals/src/components/BarChart.tsx
