+++ src/client/shared/Spaarke.Visuals/src/components/GaugeVisual.tsx
