+++ src/client/shared/Spaarke.Visuals/src/components/GradeMetricCard.tsx
