+++ src/client/shared/Spaarke.Visuals/src/components/HorizontalStackedBar.tsx
