+++ src/client/shared/Spaarke.Visuals/src/components/LineChart.tsx
