+++ src/client/shared/Spaarke.Visuals/src/components/MiniTable.tsx
