+++ src/client/shared/Spaarke.Visuals/src/components/StatusDistributionBar.tsx
