+++ src/client/shared/Spaarke.Visuals/src/utils/chartColors.ts
