+++ src/client/shared/Spaarke.Visuals/src/utils/gradeUtils.ts
