+++ src/client/shared/Spaarke.Visuals/src/utils/index.ts
@@ -1,8 +1,20 @@
 /**
  * @spaarke/visuals — utils barrel.
  *
- * Presentational helpers (chart colors, token-set colors, formatting) land
- * here in VHVU-041. Empty placeholder until then so the package typechecks.
+ * Presentational helpers (moved from the VisualHost PCF in VHVU-041):
+ *  - cardConfigResolver — resolve merged card configuration
+ *  - chartColors        — chart/status color palettes
+ *  - gradeUtils         — letter-grade + grade-color helpers
+ *  - logger             — lightweight namespaced console logger
+ *  - tokenSetColors     — Fluent token-set → color resolution
+ *  - trendAnalysis      — linear-regression slope + trend direction
+ *  - valueFormatters    — number/percentage/currency/grade formatting
  */
 
-export {};
+export * from './cardConfigResolver';
+export * from './chartColors';
+export * from './gradeUtils';
+export * from './logger';
+export * from './tokenSetColors';
+export * from './trendAnalysis';
+export * from './valueFormatters';
