+++ src/client/shared/Spaarke.Visuals/src/utils/tokenSetColors.ts
