+++ src/client/shared/Spaarke.Visuals/src/utils/valueFormatters.ts
